pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptguard"
version = "0.1.0"
description = "Reliable, structured, production-safe LLM outputs with schema validation and auto-repair"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "PromptGuard Contributors" }
]
keywords = [
    "llm",
    "ai",
    "schema",
    "validation",
    "openai",
    "anthropic",
    "pydantic",
    "structured-output",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.25",
    "tenacity>=8.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.20"]
google = ["google-generativeai>=0.4"]
testing = ["sentence-transformers>=2.0", "numpy>=1.24"]
all = [
    "promptguard[openai,anthropic,google,testing]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "ruff>=0.1",
    "promptguard[all]",
]

[project.scripts]
promptguard = "promptguard.cli.main:cli"

[project.urls]
Homepage = "https://github.com/metanoia-oss/promptguard"
Documentation = "https://github.com/metanoia-oss/promptguard"
Repository = "https://github.com/metanoia-oss/promptguard"
Issues = "https://github.com/metanoia-oss/promptguard/issues"

[tool.hatch.build.targets.wheel]
packages = ["promptguard"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["promptguard"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "openai.*",
    "anthropic.*",
    "google.generativeai.*",
    "sentence_transformers.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
